fix: TodoTaskVO -> Presentation Layer에서만 사용하도록 변경
